+++ threedigrid_builder/base/__init__.py
@@ -5,5 +5,6 @@ from .lines import *  # NOQA
 from .linestrings import *  # NOQA
 from .nodes import *  # NOQA
 from .pumps import *  # NOQA
+from .quarters import *  # NOQA
 from .settings import *  # NOQA
 from .surfaces import *  # NOQA